first egrep implemented

+++ rusterella/scripts/build.sh
@@ -3,13 +3,21 @@
 set -e
 set -u 
 
-mkdir -p temp
-cd temp
-curl -L https://github.com/upx/upx/releases/download/v3.96/upx-3.96-amd64_linux.tar.xz | tar xJf -
-mv upx-3.96-amd64_linux/upx .
-cd -
+CURFILE=$( readlink -f $0 )
+CURDIR=${CURFILE%/*}
+. ${CURDIR}/lib.sh
 
-cargo build --release
-strip target/release/rusterella
-./temp/upx target/release/rusterella
-du -sh target/release/rusterella
+
+
+if [ ! -f ./temp/upx ]; then
+  mkdir -p temp
+  cd temp
+  curl -L https://github.com/upx/upx/releases/download/v3.96/upx-3.96-amd64_linux.tar.xz | tar xJf -
+	mv upx-3.96-amd64_linux/upx .
+  cd -
+fi
+
+cargo build --profile ${PROFILE}
+strip target/${OUTPUT_PROFILE}/rusterella
+./temp/upx target/${OUTPUT_PROFILE}/rusterella
+du -sh target/${OUTPUT_PROFILE}/rusterella
